# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Audiobook API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProgress {
  chapterIndex: number;
  paragraphIndex: number;
  completedChapters: number[];
  lastUpdated: string;
}

export interface SaveProgressRequest {
  chapterIndex: number;
  paragraphIndex: number;
  completedChapters: number[];
}

export interface Bookmark {
  id: number;
  chapterIndex: number;
  paragraphIndex: number;
  chapterTitle: string;
  excerpt: string;
  createdAt: string;
}

export interface AddBookmarkRequest {
  chapterIndex: number;
  paragraphIndex: number;
  chapterTitle: string;
  excerpt: string;
}

export interface DeleteResponse {
  success: boolean;
}

export interface UserSettings {
  playbackRate: number;
  voiceURI?: string;
  theme: string;
  fontSize: string;
  autoScroll: boolean;
  highlightWords: boolean;
  pitch: number;
  volume: number;
}