# lib/api-zod/src/generated/types/addBookmarkRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Audiobook API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AddBookmarkRequest {
  chapterIndex: number;
  paragraphIndex: number;
  chapterTitle: string;
  excerpt: string;
}