# lib/api-zod/src/generated/types/bookmark.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Audiobook API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Bookmark {
  id: number;
  chapterIndex: number;
  paragraphIndex: number;
  chapterTitle: string;
  excerpt: string;
  createdAt: string;
}