# lib/api-zod/src/generated/types/saveProgressRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Audiobook API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SaveProgressRequest {
  chapterIndex: number;
  paragraphIndex: number;
  completedChapters: number[];
}