# lib/api-zod/src/generated/types/userSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Audiobook API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserSettings {
  playbackRate: number;
  voiceURI?: string;
  theme: string;
  fontSize: string;
  autoScroll: boolean;
  highlightWords: boolean;
  pitch: number;
  volume: number;
}